mm: export-GPL get_pageblock_migratetype
The macro include/linux/mmzone.h:get_pageblock_migratetype() uses
the symbol get_pfnblock_flags_mask, which is not exported. That macro is
used within the kmem and page_ref trace events to trace the migrate type
of the target page.

Exporting this symbol to GPL modules allows GPL kernel tracers to be
implemented as modules.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Cc: Andrew Morton <akpm@linux-foundation.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: linux-mm@kvack.org