writeback: tracing: pass global_wb_domain as tracepoint parameter
The symbol global_wb_domain is not GPL-exported to modules. In order
to allow kernel tracers implemented as GPL modules to access the
global writeback domain dirty limit, as used within the
global_dirty_state and balance_dirty_pages trace events, pass
a pointer to the global writeback domain as a new parameter for
those tracepoints.

In order to keep the number of parameters passed to
balance_dirty_pages below 13, move the declaration of
struct dirty_throttle_control to linux/writeback.h so it
can be used within the balance_dirty_pages tracepoint. This
allows replacing the "thread", "bg_thresh", "dirty", "bdi_thresh",
and "bdi_dirty" parameters by a single pointer to that structure.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Cc: Alexei Starovoitov <ast@kernel.org>
Cc: Daniel Borkmann <daniel@iogearbox.net>
Cc: Steven Rostedt <rostedt@goodmis.org>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: bpf@vger.kernel.org
Cc: netdev@vger.kernel.org