sched: Introduce per memory space current virtual cpu id
This feature allows the scheduler to expose a current virtual cpu id
to user-space. This virtual cpu id is within the possible cpus range,
and is temporarily (and uniquely) assigned while threads are actively
running within a memory space. If a memory space has fewer threads than
cores, or is limited to run on few cores concurrently through sched
affinity or cgroup cpusets, the virtual cpu ids will be values close
to 0, thus allowing efficient use of user-space memory for per-cpu
data structures.

This feature is meant to be exposed by a new rseq thread area field.

Implemented optimizations:

1) On context switch between threads belonging to the same memory space,
   transfer the mm_vcpu_id from prev to next without any atomic ops.
   This takes care of use-cases involving frequent context switch
   between threads belonging to the same memory space.

2) Threads belonging to memory space with single user (mm_users==1) can
   be assigned mm_vcpu_id=0 without any atomic operations on the
   scheduler fast-path. When a memory space goes from single to
   multi-threaded, lazily allocate the vcpu_id 0 in the mm vcpu mask.
   This takes care of all single-threaded use-cases involving context
   switching between threads belonging to different memory spaces.

3) Introduce a per runqueue cache containing { mm, vcpu_id } entries.
   Keep track of the recently allocated vcpu_id for each mm rather than
   freeing them immediately. This eliminates most atomic ops when
   context switching back and forth between threads belonging to
   different memory spaces in multi-threaded scenarios (many processes,
   each with many threads).

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>