ARM: Prevent speculated memory accesses across DACR load (v2)
Observed issue
==============

On a Wandboard i.MX6 Quad Board revD1 (Cortex-A9 r2p10), spurious hangs
have been experienced in sys_futex in a few scenarios with liburcu
rwlock stress-tests [1]. The test run in a loop is:

for a in $(seq 1 100000); do echo $a; tests/benchmark/.libs/test_rwlock 2 2 1 || exit 1; done

This runs a 1 second test with 2 reader and 2 writer threads.

The hangs show up in the following scenarios:

- glibc nptl rwlock,
- glibc nptl pthread_join.

In all cases a FUTEX_WAIT is stuck waiting for a state change that can
be validated to be present in user memory by inspecting the hung process
with a debugger.

This can be reproduced with a Debian bullseye config-5.18.0-0.bpo.1-armmp
configuration. The config-5.18.0-0.bpo.1-armmp configuration applied to
a vanilla 5.18.2 kernel tree reproduces this issue when compiled with
gcc version 10.2.1 20210110 (Debian 10.2.1-6).

Note that this does not appear to reproduce as soon as the configuration
varies siginificantly enough to change the kernel code generated for
futex wait:

- Compiling the kernel with gcc version 7.5.0 (Ubuntu/Linaro
  7.5.0-3ubuntu1~18.04) implements "get_current()" with an explicit
  "mrc p15, 0, %0, c13, c0, 3" rather than __builtin_thread_pointer(),
  which introduces an mcr instruction immediately after the DACR load,
  which appear to hide the issue.

- Building a imx_v6_v7_defconfig kernel with gcc 10.2.1 also hide the
  issue. This appears to be caused by the fact that this configuration
  targets __LINUX_ARM_ARCH__ = 6 rather than __LINUX_ARM_ARCH__ = 7.
  The effect of this is that isb() is implemented as
  "mcr p15, 0, rX, c7, c5, 4" rather than the "isb" instruction. The
  isb() immediately after DACR load, when implemented as a mcr
  instruction, appears to hide the issue.

Cause
=====

Note about the following discussion: these hypotheses should really be
validated by ARM staff with knowledge of the architecture internals.

The workaround for errata 754322 states:

        ARM errata: possible faulty MMU translations following an ASID switch

          This option enables the workaround for the 754322 Cortex-A9 (r2p*,
          r3p*) erratum. A speculative memory access may cause a page table walk
          which starts prior to an ASID switch but completes afterwards. This
          can populate the micro-TLB with a stale entry which may be hit with
          the new ASID. This workaround places two dsb instructions in the mm
          switching code so that no page table walks can cross the ASID switch.

When inspecting the generated assembly near the get_user() within
futex_wait_setup, we notice that set_domain() is used to allow accessing
user-space memory. set_domain is a DACR load, which is performed with a
mcr instruction.

Considering the nature of errata 754322 which is caused by a page table
walk which starts prior to the ASID switch but complete afterwards, the
hypothesis here is that similar issues may affect DACR load as well. If
a page table walk for a speculated memory access to user-space memory is
done with the wrong DACR state, this speculated access may use a stale
entry from the micro-TLB, and fail to properly access the userspace
memory.

Along this line of thinking, there may be cases where speculated memory
accesses to kernel memory performed by user-space immediately after
returning to user-space could succeed when they should fail, but this
has not been validated here. This could have security implications.

Solution
========

For configurations enabling the workaround for errata 754322
(CONFIG_ARM_ERRATA_754322=y), add a Data Synchronization Barrier (dsb)
instruction prior to:

- Domain Access Control Register (DACR) load:
             mcr     p15, 0, rX, c3, c0, 0

Assuming the page table walk speculation errata may also apply to TTBR
load and TTBCR load, to take a safe approach, also add "dsb" prior to:

- Translation Table Base Register (TTBR) load:
             mcr     p15, 0, rX, c2, c0, 0
- Translation Table Base Control Register (TTBCR) load:
             mcr     p15, 0, rX, c2, c0, 2

Known drawbacks
===============

Adding a "dsb" instructions each time the DACR, TTBR and TTBCR are
loaded slows down performance.

Link: https://liburcu.org [1]
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
---
Changes since v1:
- Use "dsb" directly from assembler.