rseq: Validate initialization of rseq_cs field
The RSEQ registration requires that userspace initialized the rseq_cs
field to NULL before registration.

Validate this with an explicit check which returns -EINVAL in case of
failure rather than just kill the application with SIGSEGV. This should
make it easier for library implementers to diagnose issues related to
registration of uninitialized struct rseq areas.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>